weatherapp/serializers.py
from rest_framework import serializers

from weatherapp.models import WeatherData


class WeatherDataSerializer(serializers.ModelSerializer):
    class Meta:
        model = WeatherData
        fields = '__all__'
