requirements.txt
amqp==5.0.6
APScheduler==3.8.0
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-apscheduler==0.6.0
djangorestframework==3.12.4
idna==3.2
kombu==5.1.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzlocal==2.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
